--- a/src/components/study/HomeDashBoard.js
+++ b/src/components/study/HomeDashBoard.js
@@ -315,4 +315,4 @@ const HomeDashBoard = () => {
         </div>
     );
 };
-export default HomeDashBoard;
+export default React.memo(HomeDashBoard);