### PR 타입

-에러 해결
-최적화

### 반영 브랜치

ex) feat/optimization -> master

### 변경 사항

로그인 한 다음 바로 나타나는 메인페이지에서 key누락 에러 해결
최적화 >> 헤더의 로고 클릭 시

최적화 전 렌더링 속도 : 6.6ms
최적화 후 렌더링 속도 : 4.1ms
최적화 전 HomeDashBoard 컴포넌트 렌더링 소요시간 0.9ms of 1.5ms
최적화 후 HomeDashBoard 컴포넌트 렌더링 소요시간 0.3ms of 0.6ms

### 테스트 결과

최적화 전 
![image](https://github.com/Hanium2023-WeB/starD-frontend/assets/80142915/e0d1e2cd-30d7-43a7-8454-6496e10abcbb)

최적화 후
![image](https://github.com/Hanium2023-WeB/starD-frontend/assets/80142915/035d71e8-ff39-475b-8852-709dd921f46d)
![image](https://github.com/Hanium2023-WeB/starD-frontend/assets/80142915/3b21edec-ad1e-4544-9dcd-268cea5d92f1)